requirements.txt
mesa>=3.0.0
numpy>=1.26.0
streamlit>=1.40.0
langchain-google-genai>=2.0.0  # The library to talk to Gemini
python-dotenv>=1.0.0